restructured tests a bit
